made ignition delay calculation more robust for species targets
For ignition delay targets of max species mole fractions, finding peaks could
lead to an error if the species continued to grow after ignition (e.g., for
a pressure-rise shock tube case...). Now, if there is no peak found for the max
mass fraction (because the peak would be the final value), the processing code
falls back on finding the peak derivative, which should still work.